requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-elasticsearch-dsl==7.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.3.1
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PySocks==1.7.1
pytest==7.4.1
pytest-base-url==2.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-html==4.0.0
pytest-metadata==3.0.0
pytest-selenium==4.0.1
pytest-variables==3.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tenacity==8.2.3
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
